Rename solver.py module to _solver.py
